Remove side effects from inference rules
Natural deduction was reimplemented in a way that the context is now
part of the judgment, instead of being a "side effect". For that, a new
subclass of Judgment was created: NaturalSequent. It is composed of a
context (a set of named formulas) in the antecedent, and a single
formula in the right. This reimplementation allowed the prover to be
greatly simplified.