Implement mutable sequent
A type class "Cedent" is used, in order to make the Sequent generic
with respect to the collection used for its cedents. But this makes
Sequent a bit complicated. Maybe it is better to simply implement two
mutable sequents: one with cedents of type Seq and one with cedents of
type Seq.