Add OverTermination condition
The condition holds when the length of the last chain of "only A2"
executions gets greater or equal to the height of the resulting proof.

There is still a bug in the condition, probably in the computation of
the height. Please don't merge this branch into the main repository
until it's fixed.