Fix UnifyingResolution & SPASS parser
The computations by Unifying Resolution now almost match what SPASS 
dictates we should see. We have more antecedents becuase we use more
than just the clauses specified; we use the whole line. Need to 
investigate if this is a concern. Also, because this change was rather
involved (or at least took a lot of work to ensure this was happening)
there are some unfinished lines in these new modifications, and a lot
of println statements that need to be removed.