Add MRR class
Currently added what I hope is a temporary class: UnifyingResolutionMRR,
meant to handle MRR lines. It's the same as UnifyingResolution, but with
one change to the output. This could very well imply a bug with
UnifyingResolution, but as it stands, using both of these classes the 
code can parse everything except the last line (which is an MRR line 
with three references, and both classes (and the unifying class) only
support up to two from the looks of it.