Add code for testing
Adds a small parser that parses output of Skeptik so that it can be fed
back in, and used. Intending to use to allow for some variance in tests
(parse the output, and if it's the same as the one that expecting, 
modulo variable renaming, then the test will pass). Should make tests
work on all systems, not just mine (when the tests are converted).