Implement draft version of ancestor-based idea
It does something "nearly" reasonable, but there are some implementation
bugs: e.g. if it decides (P(a) |= ) on some step, on the next step it can fail
to ensure trueness of (P(a) |= P(b)).

Code is very dirty. If you're trying to understand it -- better wait for new
commits: I'll certainly clean this up and add some documentation with comms.